requirements.txt
# Core ML/DL Libraries
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
peft>=0.7.0  # For LoRA fine-tuning
accelerate>=0.24.0

# LangGraph and LangChain
langgraph>=0.0.20
langchain>=0.1.0
langchain-core>=0.1.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Logging and Monitoring
loguru>=0.7.0
colorama>=0.4.6
rich>=13.0.0  # For beautiful CLI

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Visualization (Bonus)
matplotlib>=3.7.0
seaborn>=0.12.0
plotext>=5.2.0  # CLI-based plotting

# Development
jupyter>=1.0.0
ipython>=8.12.0
tqdm>=4.65.0

# Testing (Optional but recommended)
pytest>=7.4.0
pytest-cov>=4.1.0